Numbers have life; they're not just symbols on paper

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -7,7 +7,7 @@
 
 void print_number(int n)
 {
-	unsigned long int number = n;
+	unsigned int number = n;
 
 	if (n < 0)
 	{
